Initial repackaging to Python pip package
