+++ photobatch/Config.ini
@@ -1,0 +1,33 @@
+[Filepath]
+file_list_path = C:\
+event_list_path = C:\
+output_path = C:\
+
+[Event_Window]
+event_prior = 2.00
+event_follow = 5.00
+
+[ITI_Window]
+trial_start_stage = Display Image
+trial_end_stage = Reward Collected Start ITI,Start ITI,Start correction ITI
+iti_prior_trial = 1.00
+center_z = whole
+center_method = median
+
+[Photometry_Processing]
+filter_frequency = 10
+
+[Output]
+create_simplez = 1
+create_timedz = 0
+create_simplep = 0
+create_timedp = 0
+create_simplef = 0
+create_timedf = 0
+create_raw = 0
+create_summaryz = 0
+create_summaryf = 0
+create_summaryp = 0
+
+[Filter]
+exclusion_list = Remove Stimulus Duration,Blank Image Touched,Right Screen Touches,Left Screen Touches,Centre Screen Touches
