+++ requirements.txt
@@ -1,0 +1,5 @@
+PySide6
+numpy
+scipy
+pandas
+h5py
